+++ extensions/visual-studio-code-project-manager/src/search-project-manager-projects.tsx
@@ -4,7 +4,7 @@ import { existsSync, lstatSync, readFileSync } from "fs";
 import { homedir } from "os";
 import config from "parse-git-config";
 import { dirname } from "path";
-import { useState, ReactElement, Fragment } from "react";
+import { Fragment, ReactElement, useState } from "react";
 import tildify from "tildify";
 import { CachedProjectEntry, Preferences, ProjectEntry } from "./types";
 
@@ -18,14 +18,13 @@ const terminalInstalled = existsSync(terminalPath);
 
 const { vscodeApp } = preferences;
 const vscodeAppNameShort = vscodeApp?.name.replace(/^Visual Studio /, "") || "";
-const vscodeAppCLI: string = vscodeAppNameShort.replace(/[ -]+/g, "-").toLowerCase();
+const vscodeAppCLI = `${vscodeApp?.path}/Contents/Resources/app/bin/code`;
 
 const STORAGE = `${homedir()}/Library/Application Support/${vscodeAppNameShort}/User/globalStorage/alefragnani.project-manager`;
 
 const remotePrefix = "vscode-remote://";
 
-function getProjectEntries(): ProjectEntry[] {
-  const storagePath = getPreferencesPath() || STORAGE;
+function getProjectEntries(storagePath: string): ProjectEntry[] {
   const savedProjectsFile = `${storagePath}/projects.json`;
   const cachedProjectsFiles = [
     `${storagePath}/projects_cache_git.json`,
@@ -74,17 +73,25 @@ const filterProjectsByTag = (projects: ProjectEntry[], selectedTag: string): Pro
   return projects.filter((project) => (selectedTag ? project.tags?.find((tag) => tag === selectedTag) : true));
 };
 
-function getPreferencesPath(): string | undefined {
+function getProjectsLocationPath(): { path: string; error?: string } {
   const path = preferences.projectManagerDataPath;
-  if (path && existsSync(path)) {
-    const stat = lstatSync(path);
-    if (stat.isDirectory()) {
-      return path;
-    }
-    if (stat.isFile()) {
-      return dirname(path);
-    }
+  if (!path) {
+    return { path: STORAGE };
   }
+
+  if (!existsSync(path)) {
+    return { path, error: `Projects Location path does not exist: ${path}` };
+  }
+
+  const stat = lstatSync(path);
+  if (stat.isDirectory()) {
+    return { path };
+  }
+  if (stat.isFile()) {
+    return { path: dirname(path) };
+  }
+
+  return { path, error: `Projects Location path is not a directory: ${path}` };
 }
 
 function getSortedProjects(projects: ProjectEntry[]): ProjectEntry[] {
@@ -125,29 +132,41 @@ function getProjectsGroupedByTagAsElements(projectEntries: ProjectEntry[]): Reac
 
 export default function Command() {
   if (!vscodeApp) {
-    return (
-      <Detail markdown="Please configure the **Search Project Manager** Raycast extension to choose which version of Visual Studio Code to use." />
+    return ExtensionError(
+      "Please configure the **Search Project Manager** Raycast extension " +
+        "to choose which version of Visual Studio Code to use.",
     );
   }
 
-  const elements: ReactElement[] = [];
-  const projectEntries = getProjectEntries();
+  const { path: projectsLocationPath, error: projectsLocationError } = getProjectsLocationPath();
+  if (projectsLocationError) {
+    return ExtensionError(
+      "## Invalid Projects Location" +
+        "\n\n```\n" +
+        projectsLocationPath +
+        "\n```\n\n" +
+        "Please review the **Projects Location** setting in the extension configuration. " +
+        "\n\n" +
+        "This setting should only be set if Projects Location is also customized in the VS Code Project Manager extension settings.",
+    );
+  }
+
+  const projectEntries = getProjectEntries(projectsLocationPath);
   const projectTags = getProjectTags(projectEntries);
 
   const [selectedTag, setSelectedTag] = useState("");
 
   if (!projectEntries || projectEntries.length === 0) {
-    return (
-      <Detail
-        markdown="To use this extension, the Visual Studio Extension
-      [Project Manager](https://marketplace.visualstudio.com/items?itemName=alefragnani.project-manager)
-       is required and at least one project must be saved in the Project Manager."
-      />
+    return ExtensionError(
+      "To use this extension, the VS Code Extension " +
+        "[Project Manager](https://marketplace.visualstudio.com/items?itemName=alefragnani.project-manager) " +
+        "is required and at least one project must be saved in the Project Manager.",
     );
   }
 
   const sortedProjects = getSortedProjects(projectEntries);
 
+  const elements: ReactElement[] = [];
   if (preferences.groupProjectsByTag && !selectedTag) {
     // don't group if filtering
     const groupedProjects = getProjectsGroupedByTagAsElements(sortedProjects);
@@ -204,7 +223,7 @@ function ProjectListItem({ name, rootPath, tags }: ProjectEntry) {
                 title={`Open in ${vscodeApp.name} (Remote)`}
                 icon={{ fileIcon: vscodeApp.path }}
                 onAction={() => {
-                  exec(`${vscodeAppCLI} --remote ${parseRemoteURL(path)}`);
+                  exec(`"${vscodeAppCLI}" --remote ${parseRemoteURL(path)}`);
                   closeMainWindow();
                 }}
               />
@@ -252,6 +271,25 @@ function ProjectListItem({ name, rootPath, tags }: ProjectEntry) {
           </ActionPanel.Section>
           <DevelopmentActionSection />
         </ActionPanel>
+      }
+    />
+  );
+}
+
+function ExtensionError(detail: string) {
+  const { path } = getProjectsLocationPath();
+
+  return (
+    <Detail
+      markdown={detail}
+      metadata={
+        <Detail.Metadata>
+          <Detail.Metadata.Label title="VS Code App" text={vscodeApp?.name || "(unset)"} />
+          <Detail.Metadata.Label
+            title={`Projects Location${path ? "" : " (Default)"}`}
+            text={tildify(path || STORAGE)}
+          />
+        </Detail.Metadata>
       }
     />
   );