+++ extensions/open-graph/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Open Graph Changelog
 
+## [Add Browser Extension launch] - 2024-04-29
+
+- Add command that will get the url from the current open tab in your browser and open the Preview Open Graph command with this url
+
 ## [Fix image urls]
 
 - Fix the link to the images in OpenGraph