+++ extensions/simple-icons/CHANGELOG.md
@@ -1,4 +1,18 @@
-# Simple Icons Changelog
+# Brand Icons Changelog
+
+## [Updates] - 2024-05-16
+
+- Implement Raycast Cross Extension Conventions
+
+## [Updates] - 2024-04-30
+
+- Expose search function through `launchCommand`
+- Remove unused cache header
+
+## [Updates] - 2024-04-15
+
+- Rename extension name to Brand Icons
+- Update feedback links
 
 ## [New Feature & Chore] - 2024-04-09